# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Oculide Core REST API
  description: >
    Hệ thống API RESTful và WebSocket phục vụ Giám sát phòng thực hành lập trình & Gác thi bằng AI (Oculide v1).
    Được thiết kế theo tiêu chuẩn IEEE 830-1998 Specification.
  version: 1.0.0
  contact:
    name: Oculide Development Team
    email: lean80469@gmail.com
servers:
  - url: https://api.oculide.id.vn/api/v1
    description: Production API Server
  - url: http://localhost:8000/api/v1
    description: Local Development Server
paths:
...
  /auth/register:
    post:
      tags: [Authentication]
      summary: Đăng ký tài khoản người dùng mới
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegisterRequest'
      responses:
        '201':
          description: Đăng ký thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/400Error'
  /auth/login:
    post:
      tags: [Authentication]
      summary: Đăng nhập hệ thống lấy JWT Access Token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Đăng nhập thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Sai tên đăng nhập hoặc mật khẩu
  /auth/me:
    get:
      tags: [Authentication]
      summary: Lấy thông tin tài khoản đang đăng nhập
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Thông tin người dùng hiện tại
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/401Error'
  /auth/join:
    post:
      tags: [Authentication]
      summary: Sinh viên tham gia phòng nhanh bằng mã phòng và mã PIN
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JoinRoomRequest'
      responses:
        '200':
          description: Cấp token truy cập phòng thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '404':
          description: Không tìm thấy phòng hoặc phòng đã kết thúc
  /auth/profile:
    put:
      tags: [Authentication]
      summary: Cập nhật thông tin tài khoản cá nhân
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserProfileUpdateRequest'
      responses:
        '200':
          description: Cập nhật thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
  /auth/change-password:
    put:
      tags: [Authentication]
      summary: Đổi mật khẩu
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '200':
          description: Đổi mật khẩu thành công

  # ==========================================
  # ROOMS
  # ==========================================
  /rooms:
    get:
      tags: [Rooms]
      summary: Lấy danh sách các phòng học/phòng thi
      security:
        - BearerAuth: []
      parameters:
        - name: room_type
          in: query
          schema:
            type: string
            enum: [exam, classroom]
        - name: skip
          in: query
          schema:
            type: integer
            default: 0
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
      responses:
        '200':
          description: Danh sách phòng
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RoomResponse'

    post:
      tags: [Rooms]
      summary: Tạo mới một phòng thi hoặc phòng thực hành
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoomCreateRequest'
      responses:
        '201':
          description: Phòng được tạo thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoomResponse'
...
  /rooms/{room_id}:
    get:
      tags: [Rooms]
      summary: Lấy chi tiết cấu hình phòng
      security:
        - BearerAuth: []
      parameters:
        - name: room_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Chi tiết phòng
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoomResponse'

    put:
      tags: [Rooms]
      summary: Chỉnh sửa cấu hình phòng
      security:
        - BearerAuth: []
      parameters:
        - name: room_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoomUpdateRequest'
      responses:
        '200':
          description: Cập nhật thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoomResponse'

    delete:
      tags: [Rooms]
      summary: Xóa phòng
      security:
        - BearerAuth: []
      parameters:
        - name: room_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Đã xóa phòng
...
  /chat:
    post:
      tags: [Chat]
      summary: Gửi tin nhắn chat trong phòng
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatMessageRequest'
      responses:
        '201':
          description: Đã gửi tin nhắn
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatMessageResponse'
  /chat/room/{room_id}:
    get:
      tags: [Chat]
      summary: Lấy lịch sử chat trong phòng
      security:
        - BearerAuth: []
      parameters:
        - name: room_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Lịch sử tin nhắn
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    400Error:
      description: Yêu cầu không hợp lệ
    401Error:
      description: Chưa xác thực danh tính
...
  schemas:
    UserResponse:
      type: object
      properties:
        user_id:
          type: integer
        username:
          type: string
        email:
          type: string
        full_name:
          type: string
        role:
          type: string
          enum: [admin, instructor, student]
        student_id:
          type: string
          nullable: true
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
    UserRegisterRequest:
      type: object
      required: [username, password, email, full_name, role]
      properties:
        username:
          type: string
        password:
          type: string
        email:
          type: string
          format: email
        full_name:
          type: string
        role:
          type: string
          enum: [instructor, student]
        student_id:
          type: string
          nullable: true
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    TokenResponse:
      type: object
      properties:
        access_token:
          type: string
        token_type:
          type: string
          default: bearer
        user:
          $ref: '#/components/schemas/UserResponse'
    JoinRoomRequest:
      type: object
      required: [room_code]
      properties:
        room_code:
          type: string
        passcode:
          type: string
          nullable: true
        full_name:
          type: string
          nullable: true
        student_id:
          type: string
          nullable: true
    UserProfileUpdateRequest:
      type: object
      properties:
        full_name:
          type: string
        email:
          type: string
        student_id:
          type: string
    ChangePasswordRequest:
      type: object
      required: [current_password, new_password]
      properties:
        current_password:
          type: string
        new_password:
          type: string
    RoomResponse:
      type: object
      properties:
        room_id:
          type: integer
        room_code:
          type: string
        room_name:
          type: string
        room_type:
          type: string
          enum: [exam, classroom]
        instructor_id:
          type: integer
        description:
          type: string
          nullable: true
        start_time:
          type: string
          format: date-time
        end_time:
          type: string
          format: date-time
        duration_minutes:
          type: integer
        max_attempts:
          type: integer
        is_active:
          type: boolean
        livekit_room_name:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
...
    RoomCreateRequest:
      type: object
      required: [room_code, room_name, start_time, end_time, duration_minutes]
      properties:
        room_code:
          type: string
        room_name:
          type: string
        room_type:
          type: string
          default: exam
          enum: [exam, classroom]
        description:
          type: string
        start_time:
          type: string
          format: date-time
        end_time:
          type: string
          format: date-time
        duration_minutes:
          type: integer
        max_attempts:
          type: integer
          default: 1
        passcode:
          type: string
          nullable: true
    RoomUpdateRequest:
      type: object
      properties:
        room_name:
          type: string
        description:
          type: string
        start_time:
          type: string
          format: date-time
        end_time:
          type: string
          format: date-time
        duration_minutes:
          type: integer
        max_attempts:
          type: integer
        is_active:
          type: boolean
        passcode:
          type: string
...
    ChatMessageRequest:
      type: object
      required: [room_id, message_content]
      properties:
        room_id:
          type: integer
        message_content:
          type: string
        recipient_id:
          type: integer
          nullable: true
    ChatMessageResponse:
      type: object
      properties:
        message_id:
          type: integer
        room_id:
          type: integer
        sender_id:
          type: integer
        sender_name:
          type: string
        message_content:
          type: string
        sent_at:
          type: string
          format: date-time